update .env file to begin migration to absolute paths
