+++ frontend/src/components/1_initiate/modals/InitiatorVerificationModal.js
@@ -2,17 +2,17 @@ import React from 'react'
 import { connect } from 'react-redux'
 import moment from 'moment'
 import { Button } from 'antd'
-import ButtonProposeProject from '../../contractComponents/stage0/ProposeProject'
-import Map from '../shared/ModalMap'
-import ModalTemplate from './ModalTemplate.js'
+import ButtonProposeProject from 'contractComponents/stage0/ProposeProject'
+import Map from '../../shared/ModalMap'
+import ModalTemplate from 'components/modals/ModalTemplate.js'
 // import cancel from '../../images/tximages/cancel.svg'
-import txpending from '../../images/tximages/txpending.svg'
-import txconfirmed from '../../images/tximages/txconfirmed.svg'
-import txfailed from '../../images/tximages/txfailed.svg'
-import picture from '../../images/initiator/shape.png'
-import { font1 } from '../../styles/fonts'
-import { grey1, brandColor, affirmLight, cancelLight } from '../../styles/colors'
-import { web3 } from '../../utilities/blockchain'
+import txpending from 'images/tximages/txpending.svg'
+import txconfirmed from 'images/tximages/txconfirmed.svg'
+import txfailed from 'images/tximages/txfailed.svg'
+import picture from 'images/initiator/shape.png'
+import { font1 } from 'styles/fonts'
+import { grey1, brandColor, affirmLight, cancelLight } from 'styles/colors'
+import { web3 } from 'utilities/blockchain'
 
 class VerificationModal extends React.Component {
   constructor () {
