+++ frontend/src/containers/1_initiate/Initiator.js
@@ -3,19 +3,19 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import price from 'crypto-price'
-import Sidebar from '../../components/shared/Sidebar'
-import ProposeLanding from '../../components/(1) initiator/landing'
-import ProposeForm from '../../components/(1) initiator/form'
-import InitiatorWelcome from '../../components/modals/InitiatorWelcome'
-import InsufficientTokens from '../../components/modals/InsufficientTokens'
-import InitiatorVerificationModal from '../../components/modals/InitiatorVerificationModal'
-import ProjectPage from '../(2) finder/ProjectPage'
-import ipfs from '../../utilities/ipfs'
-import { getUserStatusWallet } from '../../actions/userActions'
-import { getNetworkStatus } from '../../actions/networkActions'
-import { clearTransaction } from '../../actions/transactionActions'
-import { getProject } from '../../actions/projectActions'
-import { eth, web3, dt, rr } from '../../utilities/blockchain'
+import Sidebar from 'components/shared/Sidebar'
+import ProposeLanding from 'components/1_initiate/landing'
+import ProposeForm from 'components/1_initiate/form'
+import InitiatorWelcome from 'components/1_initiate/modals/InitiatorWelcome'
+import InsufficientTokens from 'components/1_initiate/modals/InsufficientTokens'
+import InitiatorVerificationModal from 'components/1_initiate/modals/InitiatorVerificationModal'
+import ProjectPage from '../2_find/ProjectPage'
+import ipfs from 'utilities/ipfs'
+import { getUserStatusWallet } from 'actions/userActions'
+import { getNetworkStatus } from 'actions/networkActions'
+import { clearTransaction } from 'actions/transactionActions'
+import { getProject } from 'actions/projectActions'
+import { eth, web3, dt, rr } from 'utilities/blockchain'
 
 class Initiator extends React.Component {
   constructor () {
