+++ frontend/src/containers/6_resolve/Resolver.js
