+++ frontend/src/containers/Main.js
@@ -14,16 +14,16 @@ import RoleSelection from './RoleSelection'
 // import Expired from './finished/Expired'
 import Landing from './Landing'
 import Fund from './Fund'
-import Initiator from './(1) initiator/Initiator'
-import Finder from './(2) finder/Finder'
-import FinderProjectPage from './(2) finder/ProjectPage'
-import Planner from './(3) planner/Planner'
-import PlannerProjectPage from './(3) planner/ProjectPage'
-import Doer from './(4) doer/Doer'
+import Initiator from './1_initiate/Initiator'
+import Finder from './2_find/Finder'
+import FinderProjectPage from './2_find/ProjectPage'
+import Planner from './3_plan/Planner'
+import PlannerProjectPage from './3_plan/ProjectPage'
+import Doer from './4_do/Doer'
 // import DoerProjectPage from './(4) doer/ProjectPage'
-import Validator from './(5) validator/Validator'
+import Validator from './5_validate/Validator'
 // import ValidatorProjectPage from './(5) validator/ProjectPage'
-import Resolver from './(6) resolver/Resolver'
+import Resolver from './6_resolve/Resolver'
 // import ResolverProjectPage from './(6) resolver/ProjectPage'
 
 const Main = () =>
